feat: add and render FeedbackBlock component
