+++ frontend/src/features/public-form/components/FormEndPage/FormEndPage.stories.tsx
@@ -22,4 +22,5 @@ Default.args = {
     submissionId: 'mockSubmissionId',
     timeInEpochMs: 1648545566989,
   },
+  handleSubmitFeedback: (inputs) => console.log(inputs),
 }
