--- frontend/src/features/public-form/components/FormEndPage/components/EndPageBlock/index.ts
@@ -1,1 +1,0 @@
-export * from './EndPageBlock'
