+++ frontend/src/features/public-form/components/FormEndPage/components/ThankYouSvgr.tsx
